+++ packages/build-tools/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.1](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/compare/v5.3.0...v5.3.1) (2022-04-04)
+
+**Note:** Version bump only for package @bolt/build-tools
+
+
+
+
+
+# [5.3.0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/compare/v5.2.4...v5.3.0) (2022-03-28)
+
+**Note:** Version bump only for package @bolt/build-tools
+
+
+
+
+
 # [5.2.0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/compare/v5.1.1...v5.2.0) (2022-01-25)
 
 **Note:** Version bump only for package @bolt/build-tools